# contracts/iteration-12.openapi.yaml
openapi: 3.1.0
info:
  title: Bärenstark Hausservice — Iteration 12 API (Post-QA Revision)
  version: "12.1"
  description: |
    Iteration-12 API-Vertrag, post QA-Review revidiert (2026-05-04).

    Single-Source-of-Truth für alle Endpoints von IT12-S05 (Konto aus
    Buchung), IT12-S15 (Marketing-Mails inkl. Test-Send + Unsubscribe)
    und für korrigierte Specs der Bug-Stories S06/S12/S13.

    Konventionen:
    - REST/JSON über HTTPS.
    - Auth: Cookie-basiert (siehe securitySchemes).
    - Error-Format: `{ error: { code, message, field?, subcode? } }`.
    - Hard-Cap Marketing-Send IT12: 50 Empfänger pro Send-Operation
      (Vercel-Hobby-Timeout-Constraint).
    - Plain-Text Marketing-Mails only in IT12.
    - Daily-Cap Marketing: 100 Mails/Tag (Resend Free-Tier).
    - DSGVO: UWG §7 Abs. 3 (Bestandskunden), Opt-Out via Unsubscribe-Token (HMAC, stateless).
servers:
  - url: https://www.baerenstark-hausservice.app
    description: Production
  - url: http://localhost:3000
    description: Local dev
...
paths:
  /api/customer/register-from-booking:
    post:
      operationId: registerFromBooking
      summary: Konto aus einer Gast-Buchung erstellen (IT12-S05)
      description: |
        Public, Token-gated. Erstellt einen `CustomerUser` aus einer
        bestehenden Gast-Buchung. Verknüpft alle Bookings mit derselben
        E-Mail (case-insensitive) auf den neuen Account.

        Idempotenz: Wiederholtes Aufrufen mit demselben Token + bookingId
        liefert 409 wenn das Konto schon existiert. Frontend behandelt das
        als „bereits angelegt" und redirect'et zum Login.
      tags: [Customer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterFromBookingRequest'
      responses:
        '201':
          description: Konto erfolgreich erstellt, Customer-Session-Cookie gesetzt.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "customer-session=...; HttpOnly; Secure; SameSite=Lax; Path=/; Max-Age=2592000"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterFromBookingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Token ungültig oder abgelaufen (subcode INVALID_TOKEN).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Booking nicht gefunden (subcode BOOKING_NOT_FOUND).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Konto mit dieser E-Mail existiert bereits (subcode ACCOUNT_EXISTS).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiError'
                  - type: object
                    properties:
                      error:
                        type: object
                        properties:
                          subcode:
                            type: string
                            enum: [ACCOUNT_EXISTS]
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/bookings:
    post:
      operationId: createBooking
      summary: Neue Buchungsanfrage anlegen (mit Idempotency, IT12-S11)
      description: |
        Optional `Idempotency-Key` Header (UUID). Bei doppeltem Submit mit
        gleichem Key innerhalb 24h: cached Response, kein neuer Insert.
      tags: [Booking]
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          description: UUID, vom Frontend pro Submit generiert.
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Buchung angelegt.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Slot bereits vergeben (subcode SLOT_TAKEN) oder Idempotency-Conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      operationId: getAdminBookings
      summary: Alle Buchungsanfragen (Admin) — IT12-S13 verify
      description: |
        Bug-Fix in IT12-S13: Endpoint funktioniert nach `prisma migrate deploy`.
      tags: [Admin, Booking]
      security:
        - adminSession: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, CONFIRMED, REJECTED, COUNTER_PROPOSED, CANCELLED, COMPLETED]
      responses:
        '200':
          description: Liste der Buchungen.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/availability/calendar:
    get:
      operationId: getAvailabilityCalendar
      summary: Verfügbarkeits-Kalender (IT12-S03)
      description: |
        Performance-Target: p95 < 300ms bei 62-Tage-Range, 100 Bookings,
        kalter Cache. Indizes verifiziert in Build-Phase.
      tags: [Booking]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Tages-Aggregation für den Range.
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/customer/bookings:
    get:
      operationId: getCustomerBookings
      summary: Meine Buchungen (eingeloggter Customer) — IT12-S06 verify
      tags: [Customer, Booking]
      security:
        - customerSession: []
      responses:
        '200':
          description: Buchungen, gesplittet in upcoming / past.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerBookingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: |
        NextAuth-v5-Admin-Cookie (Credentials-Provider).
    customerSession:
      type: apiKey
      in: cookie
      name: customer-session
      description: |
        Custom-JWT-Cookie für Customer-Auth (siehe lib/customer-auth.ts).
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            field:
              type: string
              nullable: true
            subcode:
              type: string
              nullable: true
              description: |
                Bekannte Subcodes:
                ACCOUNT_EXISTS, INVALID_TOKEN, BOOKING_NOT_FOUND,
                SLOT_TAKEN, EMAIL_NOT_DRAFT, RECIPIENT_CAP_EXCEEDED,
                DAILY_QUOTA_EXCEEDED, RESEND_ERROR.
    RegisterFromBookingRequest:
      type: object
      required: [bookingId, confirmationToken, password]
      properties:
        bookingId:
          type: string
          minLength: 1
        confirmationToken:
          type: string
          minLength: 1
          description: |
            Signed JWT mit scope=booking-confirmation. Wird beim Erstellen
            der Buchung generiert und in der Bestätigungs-URL übergeben.
        password:
          type: string
          minLength: 12
          maxLength: 200
    RegisterFromBookingResponse:
      type: object
      required: [customerId, linkedBookingsCount]
      properties:
        customerId:
          type: string
        linkedBookingsCount:
          type: integer
          minimum: 1
...
    CustomerBookingsResponse:
      type: object
      required: [upcoming, past]
      properties:
        upcoming:
          type: array
          items: { type: object, additionalProperties: true }
        past:
          type: array
          items: { type: object, additionalProperties: true }
  responses:
    ValidationError:
      description: Validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Auth fehlt oder ungültig.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    RateLimited:
      description: Zu viele Anfragen.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }